Cleanup http and grpc packages
Update http service package structure.
Update grpc service package structure.
Update how app package works with local interface.
