+++ pkg/grpc/health.go
@@ -1,4 +1,4 @@
-package grpc
+package grpc_srv
 
 import (
 	"context"
