+++ pkg/http/not_found_handler.go
@@ -1,4 +1,4 @@
-package http
+package http_srv
 
 import (
 	"net/http"
